--- a/sqlathanor/declarative/_base_configuration_mixin.py
+++ b/sqlathanor/declarative/_base_configuration_mixin.py
@@ -583,49 +583,54 @@ class ConfigurationMixin(object):
         :raises ValueError: if ``config_set`` is not defined within ``__serialization__``
 
         """
-        if config_set and not isinstance(cls.__serialization__, (dict, OrderedDict)):
-            raise ConfigurationError('object does not use named configuration sets, '
-                                     'but config_set is not None')
-        elif not config_set and isinstance(cls.__serialization__, (dict, OrderedDict)):
-            raise ConfigurationError('configuration sets defined but config_set is empty')
+        def _build_set():
+            if config_set and not isinstance(cls.__serialization__, (dict, OrderedDict)):
+                raise ConfigurationError('%s object does not use named (%s) configuration sets, '
+                                         'but config_set is not None' % (cls, config_set))
+            elif not config_set and isinstance(cls.__serialization__, (dict, OrderedDict)):
+                raise ConfigurationError('%s configuration sets defined but config_set is empty' % cls)
 
-        if config_set and config_set not in cls.__serialization__:
-            raise ValueError('config set (%s) not found in __serialization__ '
-                             ' configuration' % config_set)
+            if config_set and config_set not in cls.__serialization__:
+                raise ValueError('%s config set (%s) not found in __serialization__ '
+                                 ' configuration' % (cls, config_set))
 
-        if config_set:
-            __serialization__ = [x for x in cls.__serialization__[config_set]]
-        else:
-            __serialization__ = [x for x in cls.__serialization__]
+            if config_set:
+                __serialization__ = [x for x in cls.__serialization__[config_set]]
+            else:
+                __serialization__ = [x for x in cls.__serialization__]
 
-        declarative_attributes = cls._get_declarative_serializable_attributes(
-            from_csv = from_csv,
-            to_csv = to_csv,
-            from_json = from_json,
-            to_json = to_json,
-            from_yaml = from_yaml,
-            to_yaml = to_yaml,
-            from_dict = from_dict,
-            to_dict = to_dict,
-            exclude_private = exclude_private
-        )
-        meta_attributes = cls._get_meta_serializable_attributes(
-            from_csv = from_csv,
-            to_csv = to_csv,
-            from_json = from_json,
-            to_json = to_json,
-            from_yaml = from_yaml,
-            to_yaml = to_yaml,
-            from_dict = from_dict,
-            to_dict = to_dict,
-            config_set = config_set
-        )
+            declarative_attributes = cls._get_declarative_serializable_attributes(
+                from_csv = from_csv,
+                to_csv = to_csv,
+                from_json = from_json,
+                to_json = to_json,
+                from_yaml = from_yaml,
+                to_yaml = to_yaml,
+                from_dict = from_dict,
+                to_dict = to_dict,
+                exclude_private = exclude_private
+            )
+            meta_attributes = cls._get_meta_serializable_attributes(
+                from_csv = from_csv,
+                to_csv = to_csv,
+                from_json = from_json,
+                to_json = to_json,
+                from_yaml = from_yaml,
+                to_yaml = to_yaml,
+                from_dict = from_dict,
+                to_dict = to_dict,
+                config_set = config_set
+            )
 
-        attributes = [x for x in meta_attributes]
-        attributes.extend([x for x in declarative_attributes
-                           if x not in attributes and x not in __serialization__])
+            attributes = [x for x in meta_attributes]
+            attributes.extend([x for x in declarative_attributes
+                               if x not in attributes and x not in __serialization__])
 
-        return attributes
+            return attributes
+
+        k = f'attributes_fc{str(from_csv)}tc{str(to_csv)}fj{str(from_json)}tj{str(to_json)}fy{str(from_yaml)}ty{str(to_yaml)}fd{str(from_dict)}td{str(to_dict)}ep{str(exclude_private)}',
+
+        return cls._heapable(k, _build_set, config_set)
 
     @classmethod
     def get_attribute_serialization_config(cls,
@@ -656,12 +661,10 @@ class ConfigurationMixin(object):
         :raises ValueError: if ``config_set`` is not defined within ``__serialization__``
 
         """
-        attributes = cls._get_attribute_configurations(config_set = config_set)
+        attributes = cls._heapable(f'config_{attribute}', lambda: cls._get_attribute_configurations(config_set = config_set), config_set)
 
         for config in attributes:
-            if config.name == attribute:
-                return config.copy()
-            if config.display_name == attribute:
+            if config.name == attribute or config.display_name == attribute:
                 return config.copy()
 
         return None
@@ -979,6 +982,8 @@ class ConfigurationMixin(object):
         else:
             cls.__serialization__ = [x for x in serialization]
 
+        cls.clear_serialization_cache()
+
     @classmethod
     def configure_serialization(cls,
                                 configs = None,
@@ -1183,6 +1188,8 @@ class ConfigurationMixin(object):
         :type config_set: :class:`str <python:str>` / :obj:`None <python:None>`
 
         """
+        cls.clear_serialization_cache()
+
         config = validate_serialization_config(configs)
         config_attributes = [x.name for x in config]
 
@@ -1421,15 +1428,18 @@ class ConfigurationMixin(object):
         :raises ValueError: if ``config_set`` is not defined within ``__serialization__``
 
         """
-        attributes = [x.copy()
-                      for x in cls.get_serialization_config(from_csv = deserialize,
-                                                            to_csv = serialize,
-                                                            config_set = config_set)]
-        for config in attributes:
-            if config.csv_sequence is None:
-                config.csv_sequence = len(attributes) + 1
+        def _build_csv():
+            attributes = [x.copy()
+                          for x in cls.get_serialization_config(from_csv=deserialize,
+                                                                to_csv=serialize,
+                                                                config_set=config_set)]
+            for config in attributes:
+                if config.csv_sequence is None:
+                    config.csv_sequence = len(attributes) + 1
 
-        return sorted(attributes, key = lambda x: (x.csv_sequence, x.name))
+            return sorted(attributes, key=lambda x: (x.csv_sequence, x.name))
+
+        return cls._heapable(f'csv_s{str(serialize)}d{str(deserialize)}', _build_csv, config_set)
 
     @classmethod
     def get_json_serialization_config(cls,
@@ -1464,9 +1474,10 @@ class ConfigurationMixin(object):
         :rtype: :class:`list <python:list>` of
           :class:`AttributeConfiguration <sqlathanor.attributes.AttributeConfiguration>`
         """
-        return [x for x in cls.get_serialization_config(from_json = deserialize,
+        return cls._heapable(f'json_s{str(serialize)}d{str(deserialize)}',
+                             lambda: [x for x in cls.get_serialization_config(from_json = deserialize,
                                                         to_json = serialize,
-                                                        config_set = config_set)]
+                                                        config_set = config_set)], config_set)
 
     @classmethod
     def get_yaml_serialization_config(cls,
@@ -1502,9 +1513,10 @@ class ConfigurationMixin(object):
           :class:`AttributeConfiguration <sqlathanor.attributes.AttributeConfiguration>`
 
         """
-        return [x for x in cls.get_serialization_config(from_yaml = deserialize,
+        return cls._heapable(f'yaml_s{str(serialize)}d{str(deserialize)}',
+                             lambda: [x for x in cls.get_serialization_config(from_yaml = deserialize,
                                                         to_yaml = serialize,
-                                                        config_set = config_set)]
+                                                        config_set = config_set)], config_set)
 
     @classmethod
     def get_dict_serialization_config(cls,
@@ -1540,6 +1552,24 @@ class ConfigurationMixin(object):
         :rtype: :class:`list <python:list>` of
           :class:`AttributeConfiguration <sqlathanor.attributes.AttributeConfiguration>`
         """
-        return [x for x in cls.get_serialization_config(from_dict = deserialize,
+        return cls._heapable(f'dict_s{str(serialize)}d{str(deserialize)}', lambda: [x for x in cls.get_serialization_config(from_dict = deserialize,
                                                         to_dict = serialize,
-                                                        config_set = config_set)]
+                                                        config_set = config_set)], config_set)
+
+    @classmethod
+    def _heapable(cls, name, value, config_set=None):
+        _heap_name = f'__serialization_cache__{name}__{"default" if config_set is None else config_set}__'
+
+        if not hasattr(cls, _heap_name):
+            _v = value()
+            if _v is not None:
+                setattr(cls, _heap_name, _v)
+
+        if hasattr(cls, _heap_name):
+            return getattr(cls, _heap_name)
+
+    @classmethod
+    def clear_serialization_cache(cls):
+        for n in [*cls.__dict__.keys()]:
+            if n.startswith('__serialization_cache__'):
+                delattr(cls, n)