by using some heap caches on classes we can massively increase the serialization/deserialize performance.

some numbers: 

## Setup

**schema** : `entity : many to many : entity2`
**objects** : 300 (100x entity + 100x relation class + 100x entity2)

## old

**mean** : 3.5 sec 
**serialization per entity** : 12ms

## new

**mean** : 450 ms
**serialization per entity** : 1.5ms


so we have a performance improvement of roughly **factor 8**.
Instead of calculating the serialisation configs for every entity again and again, we do it once per class and config_set, and store the calculated config in a heap cache on the class.

all tests are passing.

fixes #76