@insightindustry 2020-12-10
@digitalkaoz : Awesome - thanks so much for this pull request! The idea of heap caches is a great one, and I think this is definitely going to provide some massive performance improvements. I've gone and canceled the TravisCI test build and will be restarting it momentarily, and I'll also dive in and review the PR more fully very shortly. I definitely want to merge this in and release a new version with this performance improvement ASAP! Thanks again - more comments to come shortly...