@insightindustry 2020-12-10
Just as a further side note about the Travis CI tests - it turns out the "tests getting stuck in the queued state" problem is tied to Travis CI's migration of open-source projects from travis-ci.org to travis-ci.com. They've apparently got massive performance and backlog issues with running builds on their Linux environments on travis-ci.org at the moment. It's churning through all of the tests again at the moment, but it'll probably take a long time. 

I'll let the test suite run overnight, and if it doesn't finish by end-of-day tomorrow, I'll cancel the build and manually migrate things to travis-ci.com, where apparently they are not having these kind of performance / build backlog issues.