@digitalkaoz 2020-12-11
@insightindustry travis still doesnt look good, i guess a migration to github actions or the other travis infrastructure would be suitable