@insightindustry 2020-12-25
@digitalkaoz Thanks for your patience with this pull request! It took a while to successfully migrate the SQLAthanor library over to the new Travis-CI.com environment for the test matrix, but now i've been able to do that succesfully and the tests are executing properly. In the course of those tests, it looks like there are some backwards-compatability syntax issues which I mentioned in the review of the PR - if you could make those tweaks and commit, hopefully the test results will show the tests all passing and we can roll out the update to the library!