@digitalkaoz 2020-12-26
Yes i happilly will!

Insight Industry Inc. <notifications@github.com> schrieb am Fr., 25. Dez.
2020, 21:00:

> @digitalkaoz <https://github.com/digitalkaoz> Thanks for your patience
> with this pull request! It took a while to successfully migrate the
> SQLAthanor library over to the new Travis-CI.com environment for the test
> matrix, but now i've been able to do that succesfully and the tests are
> executing properly. In the course of those tests, it looks like there are
> some backwards-compatability syntax issues which I mentioned in the review
> of the PR - if you could make those tweaks and commit, hopefully the test
> results will show the tests all passing and we can roll out the update to
> the library!
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/insightindustry/sqlathanor/pull/97#issuecomment-751286181>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AACHVV6IZDIKC2XQVIRHLNLSWTVNLANCNFSM4USXOBEA>
> .
>