@digitalkaoz 2020-12-28
@insightindustry so travis is happy with every python version in the test matrix now. would love to see this merged now :)